requirements.txt
markdownify
python-dotenv
requests
